refactor(test): split e2e test cases into separate files
Inprove extensibility and maintainability by splitting the end-to-end
test cases into separate files.
Also use string_view instead of string in the AstarteMesage constructor
to pass the interface name, this way it is possible to pass a string literal
without the need to create a temporary string object (useful when testing).

Signed-off-by: Riccardo Gallo <riccardo.gallo@secomind.com>