feat(thread): use std::jthread instaed of std::thread
Introduce jthread and the new cancellation token mechanism to handle
threads termination:
- Remove the kill_reception_ atomic bool and use token.stop_requested()
instead
- Pass a weak reference of the shared stop_source to cleanly disconnect
an attached device
Make the orchestrator clean up the testcases every time one of them
is executed by leaveraging on the cancellation token management.
Finally, the introduction of an atomic bool representing the connection
state of a node allows us to remove the timeout from the is_connected
and dynamic introspection methods.

Signed-off-by: Riccardo Gallo <riccardo.gallo@secomind.com>